# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Task",
        "description": "Auth with JWT, and users can create Statements",
        "contact": {
            "email": "ashotshushanyan14@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/statements": {
            "get": {
                "tags": [
                    "Statement"
                ],
                "summary": "Get all statements",
                "description": "Returns a list of all statements",
                "operationId": "f3d3a89fdad525e9ac50fabf65ccbfcb",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "number": {
                                                "type": "string"
                                            },
                                            "date": {
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "updated_at": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "id": {
                                                "type": "integer"
                                            },
                                            "files": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "file_name": {
                                                            "type": "string"
                                                        },
                                                        "file_path": {
                                                            "type": "string"
                                                        },
                                                        "statement_id": {
                                                            "type": "integer"
                                                        },
                                                        "created_at": {
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Statement"
                ],
                "summary": "Store a new statement",
                "description": "Store a new statement and its related files.",
                "operationId": "882023a1f42af1cb8a7e18e58a7a0282",
                "requestBody": {
                    "description": "The data needed to create a new statement.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "statement_files": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "file": {
                                                    "type": "string"
                                                },
                                                "filename": {
                                                    "type": "string"
                                                },
                                                "statement_id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The new statement was created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "number": {
                                            "type": "string"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "id": {
                                            "type": "integer"
                                        },
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "file_name": {
                                                        "type": "string"
                                                    },
                                                    "file_path": {
                                                        "type": "string"
                                                    },
                                                    "statement_id": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The data provided is invalid or incomplete."
                    },
                    "500": {
                        "description": "An error occurred while processing the request."
                    }
                }
            }
        },
        "/api/statements/{id}": {
            "put": {
                "tags": [
                    "Statement"
                ],
                "summary": "Update an existing statement",
                "description": "Update an existing statement and its related files.",
                "operationId": "2a0b3a9cbbbd0694d58c733dbe6dbf6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the statement to update.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The data needed to update the statement.",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "The statement was updated successfully."
                    },
                    "404": {
                        "description": "The statement with the provided ID was not found."
                    },
                    "422": {
                        "description": "The data provided is invalid or incomplete."
                    },
                    "500": {
                        "description": "An error occurred while processing the request."
                    }
                }
            }
        },
...
    }
}